Move most of the code to gepetto-viewer.
