--- include/gepetto/gui/config-dep.hh.in
@@ -1,6 +1,0 @@
-#ifndef GEPETTO_GUI_CONFIG_DEP_HH
-# define GEPETTO_GUI_CONFIG_DEP_HH
-
-#cmakedefine01 GEPETTO_GUI_HAS_PYTHONQT
-
-#endif // GEPETTO_GUI_CONFIG_DEP_HH
