+++ src/CMakeLists.txt
@@ -79,24 +79,8 @@ IF(NOT CLIENT_ONLY)
     ${CMAKE_CURRENT_BINARY_DIR}/gepetto/viewer/corba/graphical-interfaceSK.cc
 
     ${CMAKE_BINARY_DIR}/src/gui/main.cc
-    gui/settings.cc
-    gui/windows-manager.cc
     gui/safeapplication.cc
-    gui/mainwindow.cc
-    gui/osgwidget.cc
-    gui/pick-handler.cc
-    gui/tree-item.cc
     gui/omniorbthread.cc
-    gui/bodytreewidget.cc
-    gui/ledindicator.cc
-    gui/dialog/dialogloadrobot.cc
-    gui/dialog/dialogloadenvironment.cc
-    gui/dialog/pluginmanagerdialog.cc
-    gui/shortcut-factory.cc
-    gui/selection-handler.cc
-    gui/selection-event.cc
-    gui/action-search-bar.cc
-    gui/node-action.cc
     )
 
   # Removes 5 warnings about anonymous variadic macros that were added from
@@ -104,41 +88,23 @@ IF(NOT CLIENT_ONLY)
   SET_SOURCE_FILES_PROPERTIES(graphical-interface.impl.cpp
     PROPERTIES
     COMPILE_FLAGS "-Wno-variadic-macros")
-  # }}}
 
-  # {{{ Configure PythonQt
-  IF(GEPETTO_GUI_HAS_PYTHONQT)
-    INCLUDE_DIRECTORIES(SYSTEM "${PYTHON_INCLUDE_DIR}" "${PYTHONQT_INCLUDE_DIR}")
-    SET (${PROJECT_NAME}_SOURCES
-      ${${PROJECT_NAME}_SOURCES}
-      gui/pythonwidget.cc)
-    SET (${PROJECT_NAME}_MOC
-      ${${PROJECT_NAME}_HEADERS_MOC}
-      gui/python-decorator.hh
-      )
-  ELSE (GEPETTO_GUI_HAS_PYTHONQT)
-    MESSAGE(STATUS "Skipping PythonQt settings")
-    SET (${PROJECT_NAME}_MOC
-      ${${PROJECT_NAME}_HEADERS_MOC}
-      )
-  ENDIF(GEPETTO_GUI_HAS_PYTHONQT)
+  SET (${PROJECT_NAME}_MOC
+    ${${PROJECT_NAME}_HEADERS_MOC}
+    )
   # }}}
 
   # {{{ QT config: Compile meta-objects and executable
-  SET (${PROJECT_NAME}_HEADERS
-    ${${PROJECT_NAME}_HEADERS_MOC}
-    ${${PROJECT_NAME}_HEADERS_NO_MOC}
+  SET (${PROJECT_NAME}_MOC
+    gui/safeapplication.hh
+    gui/omniorbthread.hh
     )
 
   IF(USE_QT4)
     QT4_WRAP_CPP(${PROJECT_NAME}_HEADERS_MOCED ${${PROJECT_NAME}_MOC})
-    QT4_WRAP_UI(${PROJECT_NAME}_FORMS_HEADERS ${${PROJECT_NAME}_FORMS})
-    QT4_ADD_RESOURCES(${PROJECT_NAME}_RESOURCES_RCC ${${PROJECT_NAME}_RESOURCES})
     INCLUDE(${QT_USE_FILE})
   ELSE(USE_QT4)
     QT5_WRAP_CPP(${PROJECT_NAME}_HEADERS_MOCED ${${PROJECT_NAME}_MOC})
-    QT5_WRAP_UI(${PROJECT_NAME}_FORMS_HEADERS ${${PROJECT_NAME}_FORMS})
-    QT5_ADD_RESOURCES(${PROJECT_NAME}_RESOURCES_RCC ${${PROJECT_NAME}_RESOURCES})
   ENDIF(USE_QT4)
 
   ADD_DEFINITIONS(${QT_DEFINITIONS})
@@ -147,9 +113,7 @@ IF(NOT CLIENT_ONLY)
   # {{{ Compile, link and install
   ADD_EXECUTABLE(gepetto-gui ${${PROJECT_NAME}_SOURCES}
     ${${PROJECT_NAME}_HEADERS_MOCED}
-    ${${PROJECT_NAME}_HEADERS_NO_MOC}
-    ${${PROJECT_NAME}_FORMS_HEADERS}
-    ${${PROJECT_NAME}_RESOURCES_RCC}
+    ${${PROJECT_NAME}_HEADERS}
     )
 
   ADD_DEPENDENCIES (gepetto-gui generate_idl_cpp)
@@ -157,17 +121,9 @@ IF(NOT CLIENT_ONLY)
 
   TARGET_LINK_LIBRARIES(gepetto-gui ${QT_LIBRARIES} ${Boost_LIBRARIES})
   PKG_CONFIG_USE_DEPENDENCY(gepetto-gui gepetto-viewer)
-  PKG_CONFIG_USE_DEPENDENCY(gepetto-gui openscenegraph)
-  PKG_CONFIG_USE_DEPENDENCY(gepetto-gui openthreads)
-  PKG_CONFIG_USE_DEPENDENCY(gepetto-gui ${OSGQT_LIB_NAME})
   PKG_CONFIG_USE_DEPENDENCY(gepetto-gui omniORB4)
-  IF (GEPETTO_GUI_HAS_PYTHONQT)
-    TARGET_LINK_LIBRARIES(gepetto-gui ${PYTHONQT_LIBRARIES})
-  ENDIF (GEPETTO_GUI_HAS_PYTHONQT)
 
   INSTALL(TARGETS gepetto-gui DESTINATION bin)
-
-  #PKG_CONFIG_USE_DEPENDENCY(${LIBRARY_NAME} urdfdom)
 
   # }}}
