+++ src/gui/omniorbthread.cc
@@ -14,7 +14,7 @@
 // received a copy of the GNU Lesser General Public License along with
 // gepetto-viewer-corba. If not, see <http://www.gnu.org/licenses/>.
 
-#include "gepetto/gui/omniorb/omniorbthread.hh"
+#include "gui/omniorbthread.hh"
 
 #include <gepetto/viewer/corba/server.hh>
 #include <QDebug>
@@ -22,93 +22,89 @@
 #include <gepetto/gui/mainwindow.hh>
 #include <gepetto/gui/plugin-interface.hh>
 
-namespace gepetto {
-  namespace gui {
-    ServerProcess::ServerProcess()
-      : initDone_ ()
-    {
-      initDone_.lock();
-    }
+ServerProcess::ServerProcess()
+  : initDone_ ()
+{
+  initDone_.lock();
+}
 
-    void ServerProcess::waitForInitDone()
-    {
-      initDone_.lock();
-      initDone_.unlock();
-    }
+void ServerProcess::waitForInitDone()
+{
+  initDone_.lock();
+  initDone_.unlock();
+}
 
-    void ServerProcess::init()
-    {
-      initDone_.unlock();
-    }
+void ServerProcess::init()
+{
+  initDone_.unlock();
+}
 
-    ViewerServerProcess::ViewerServerProcess (graphics::corbaServer::Server *server)
-      : server_ (server)
-    {}
+ViewerServerProcess::ViewerServerProcess (graphics::corbaServer::Server *server)
+  : server_ (server)
+{}
 
-    ViewerServerProcess::~ViewerServerProcess()
-    {
-      delete server_;
-    }
+ViewerServerProcess::~ViewerServerProcess()
+{
+  delete server_;
+}
 
-    void ViewerServerProcess::init()
-    {
-      server_->startCorbaServer ();
-      emit done ();
-      ServerProcess::init();
-    }
+void ViewerServerProcess::init()
+{
+  server_->startCorbaServer ();
+  emit done ();
+  ServerProcess::init();
+}
 
-    void ViewerServerProcess::processRequest(bool loop)
-    {
-      server_->processRequest (loop);
-      emit done();
-    }
+void ViewerServerProcess::processRequest(bool loop)
+{
+  server_->processRequest (loop);
+  emit done();
+}
 
-    CorbaServer::CorbaServer (ServerProcess* process) :
-      QObject (), control_ (process), worker_ (), timerId_ (-1), interval_ (100)
-    {
-      connect (this, SIGNAL (process(bool)), control_, SLOT (processRequest (bool)));
-      connect (control_, SIGNAL (done()), this, SLOT (processed()));
-      connect (&worker_, SIGNAL (started()), control_, SLOT (init()));
-      control_->moveToThread(&worker_);
-    }
+CorbaServer::CorbaServer (ServerProcess* process) :
+  QObject (), control_ (process), worker_ (), timerId_ (-1), interval_ (100)
+{
+  connect (this, SIGNAL (process(bool)), control_, SLOT (processRequest (bool)));
+  connect (control_, SIGNAL (done()), this, SLOT (processed()));
+  connect (&worker_, SIGNAL (started()), control_, SLOT (init()));
+  control_->moveToThread(&worker_);
+}
 
-    CorbaServer::~CorbaServer()
-    {
-      if (control_)
-        delete control_;
-    }
+CorbaServer::~CorbaServer()
+{
+  if (control_)
+    delete control_;
+}
 
-    void CorbaServer::wait ()
-    {
-      worker_.quit ();
-      worker_.wait(200);
-      if (worker_.isRunning()) {
-        worker_.terminate();
-        worker_.wait();
-      }
-    }
+void CorbaServer::wait ()
+{
+  worker_.quit ();
+  worker_.wait(200);
+  if (worker_.isRunning()) {
+    worker_.terminate();
+    worker_.wait();
+  }
+}
 
-    void CorbaServer::waitForInitDone()
-    {
-      control_->waitForInitDone();
-    }
+void CorbaServer::waitForInitDone()
+{
+  control_->waitForInitDone();
+}
 
-    void CorbaServer::start()
-    {
-      worker_.start();
-    }
+void CorbaServer::start()
+{
+  worker_.start();
+}
 
-    void CorbaServer::timerEvent(QTimerEvent* event)
-    {
-      Q_UNUSED (event);
-      assert (event->timerId () == timerId_);
-      emit process (false);
-      killTimer(timerId_);
-    }
+void CorbaServer::timerEvent(QTimerEvent* event)
+{
+  Q_UNUSED (event);
+  assert (event->timerId () == timerId_);
+  emit process (false);
+  killTimer(timerId_);
+}
 
-    void CorbaServer::processed()
-    {
-      timerId_ = startTimer(interval_);
-    }
-  } // namespace gui
-} // namespace gepetto
+void CorbaServer::processed()
+{
+  timerId_ = startTimer(interval_);
+}
