+++ tests/CMakeLists.txt
@@ -37,9 +37,6 @@ MACRO(ADD_TESTCASE NAME GENERATED)
   ENDIF()
 
   TARGET_LINK_LIBRARIES(${NAME} ${QT_LIBRARIES})
-  PKG_CONFIG_USE_DEPENDENCY(${NAME} openscenegraph)
-  PKG_CONFIG_USE_DEPENDENCY(${NAME} openthreads)
-  PKG_CONFIG_USE_DEPENDENCY(${NAME} ${OSGQT_LIB_NAME})
   PKG_CONFIG_USE_DEPENDENCY(${NAME} gepetto-viewer)
 
 ENDMACRO(ADD_TESTCASE)
